+++ tests/blackbox/test_v1api_settings.py
@@ -149,6 +149,25 @@ def test_declare_conflict(size_limit_setting, app_client):
     assert res.status_code == 409
 
 
+def test_declare_type_upgrade(size_limit_setting, app_client):
+    upgraded_type = 'float'
+    updated_setting = {
+        'name': 'size_limit',
+        'configurable_features': ['user', 'theme'],
+        'type': upgraded_type,
+        'default_value': 200,
+        'metadata': {'testing': True}
+    }
+
+    res = app_client.put('api/v1/settings/declare', data=json.dumps(updated_setting))
+    res.raise_for_status()
+    assert res.json() == {'changed': ['type'], 'created': False, 'incomplete': {}}
+
+    res = app_client.get('api/v1/settings/size_limit')
+    res.raise_for_status()
+    assert res.json() == updated_setting
+
+
 def test_declare_incomplete(size_limit_setting, app_client):
     res = app_client.put('api/v1/settings/declare', data=json.dumps({
         'name': 'size_limit',