+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Heksher Changelog
 ## unreleased
+### Added
+* Added the ability of type upgrade via the 'declare' route.
 ## 0.2.3
 ### Added
 * `patch` method for rules. Existing rules values can be patched using the new method.